% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_table.R
\name{correlation_table}
\alias{correlation_table}
\title{Calculates correlation coefficients}
\usage{
correlation_table(data, docname = "Correlation_table", save_to_doc = FALSE)
}
\arguments{
\item{data}{A data frame or matrix containing the all variables for correlation.}

\item{docname}{The name of the output MS Word document (if saving to file).}

\item{save_to_doc}{Logical, whether to save the correlation table to a Word document.}
}
\value{
A correlation table as a data frame.
}
\description{
This function computes pairwise Pearson correlation coefficients for a given dataset.
It also computes confidence intervals for the correlations and outputs the
results as a table. The table can be saved as a Word document.
}
\examples{
# Example data
corr_data <- data.frame(
  Var1 = c(1, 2, 3, 4, 5),
  Var2 = c(2, 4, 6, 8, 10),
  Var3 = c(5, 4, 3, 2, 1)
)

# Generate the correlation table
correlation_table(data = corr_data, save_to_doc = FALSE)

}
