% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_tp.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing data}
\usage{
impute_missing(data, impute_cols, sample_size = "n_in_arm")
}
\arguments{
\item{data}{A data frame containing the input data to run imputations for missing data rows.}

\item{impute_cols}{A character vector specifying the column names in `data` to impute missing values for.}

\item{sample_size}{A string specifying the column name in `data` that contains the sample size values. Default is `"n_in_arm"`.}
}
\value{
A data frame with imputed values for the specified columns and new columns appended for each imputed column,
storing the total sample sizes used in the imputation process.
The new columns are named as `"{col}_sample"`.
}
\description{
This function imputes missing values as proxies for calculations in specified columns by
replacing them with weighted means,based on the corresponding sample sizes.
It also creates new columns to store the sample sizes used for the imputation.
}
\details{
- For each specified column, the function calculates a weighted mean of non-missing values, using the sample size as weights.
- Missing values are replaced with the weighted mean.
- A new column is created for each imputed column, storing the total sample size used for imputation.
- If there are no valid rows (all values are `NA`), the function assigns `NA` to the imputed values and sample size column.
}
\examples{
# Example dataset
data <- data.frame(
  week1 = c(NA, 5, 7, 8, NA),
  week2 = c(3, NA, 6, NA, 4),
  n_in_arm = c(10, 15, 20, 25, 30)
)

impute_cols <- c("week1", "week2")

# Run imputation
imputed_data <- impute_missing(data, impute_cols = impute_cols, sample_size = "n_in_arm")

# View results
print(imputed_data)

}
